requirements.txt
flask==2.3.3
flask-restful==0.3.10
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
langchain-groq==0.1.7
langchain-huggingface==0.0.2
langchain-chroma==0.0.3
langchain-community==0.0.19
langchain-text-splitters==0.0.1
langchain-core==0.1.28
langgraph==0.1.18
typing-extensions==4.9.0